requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==24.1.1
certifi==2023.11.17
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==45.0.4
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mongoengine==0.27.0
motor==3.7.1
mypy-extensions==1.0.0
orjson==3.9.12
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
PyJWT==2.8.0
pymongo==4.13.2
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
six==1.17.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
